Added base dataset class, removed default random resized crop during training
