+++ config/mnist.yaml
@@ -7,8 +7,8 @@ MODEL:
   N_FEATURES: 32
 OPTIM:
   BATCH_SIZE: 32
-  BASE_LR: 0.0001
-  DISC_LR: 0.00001
+  BASE_LR: 0.00005
+  DISC_LR: 0.000005
   EPOCH: 2
   SCHEDULER:
     TYPE: 'OneCycle'
@@ -21,4 +21,6 @@ LOSS:
   REC_LOSS: 'BCE'
   REC_LOSS_COEFF: 1.
   PRIOR_LOSS_COEFF: 0.1
-  ADV_LOSS_COEFF: 0.5
+  ADV_LOSS_COEFF: 0.3
+DEBUG:
+  DETECT_THRESH: 0.5
