+++ data/transforms.py
@@ -4,14 +4,13 @@ from torchvision import transforms
 transforms_map = {
 	"HorizontalFlip": transforms.RandomHorizontalFlip,
 	"VerticalFlip": transforms.RandomVerticalFlip,
-	"Rotation": transforms.RandomRotation
+	"Rotation": transforms.RandomRotation,
+	"Resize": transforms.Resize,
+	"RandomResizedCrop": transforms.RandomResizedCrop
 }
 
 def build_transforms(transforms_list, input_size, train):
-	if train:
-		img_transforms = [transforms.ToPILImage(), transforms.RandomResizedCrop(input_size)]
-	else:
-		img_transforms = [transforms.ToPILImage(), transforms.Resize(input_size)]
+	img_transforms = [transforms.ToPILImage()]
 	for (transf, transf_args) in transforms_list:
 		if transf_args is not None:
 			img_transforms.append(transforms_map[transf](**transf_args))
