+++ test.py
@@ -59,7 +59,9 @@ def main():
     dataset = MNISTDataset(
         data_dir=args.dataset_dir,
         split='test',
-        input_size=config.DATASET.INPUT_SIZE)
+        input_size=config.DATASET.INPUT_SIZE,
+        transforms=[
+            ("Resize", {"size": config.DATASET.INPUT_SIZE})])
     random.shuffle(dataset.data)
     dataset.data = dataset.data[:args.n_plot*args.batch_size]
     test_loader = DataLoader(dataset, batch_size=args.batch_size, shuffle=False)
